major restructing complete, most things are broken
