+++ go.mod
@@ -3,13 +3,13 @@ module github.com/maxmcd/bramble
 go 1.16
 
 require (
+	github.com/Azure/go-ansiterm v0.0.0-20170929234023-d6e3b3328b78 // indirect
 	github.com/BurntSushi/toml v0.3.1
 	github.com/bmatcuk/doublestar/v4 v4.0.1
 	github.com/certifi/gocertifi v0.0.0-20200922220541-2c3bb06c6054
 	github.com/containerd/console v1.0.0
 	github.com/creack/pty v1.1.11
 	github.com/docker/docker v1.4.2-0.20191101170500-ac7306503d23
-	github.com/fsouza/go-dockerclient v1.6.5
 	github.com/go-git/go-git/v5 v5.3.0
 	github.com/google/go-cmp v0.5.5 // indirect
 	github.com/jaguilar/vt100 v0.0.0-20201024211400-81de19cb81a4
@@ -23,9 +23,9 @@ require (
 	github.com/tonistiigi/units v0.0.0-20180711220420-6950e57a87ea
 	go.starlark.net v0.0.0-20200901195727-6e684ef5eeee
 	go.uber.org/zap v1.10.0
-	golang.org/x/sync v0.0.0-20201020160332-67f06af15bc9 // indirect
 	golang.org/x/sys v0.0.0-20210324051608-47abb6519492
 	golang.org/x/time v0.0.0-20200630173020-3af7569d3a1e
+	gotest.tools v2.2.0+incompatible // indirect
 )
 
 replace (
