+++ pkg/dstruct/data_structures.go
@@ -1,4 +1,4 @@
-package bramble
+package dstruct
 
 import (
 	"fmt"
@@ -7,32 +7,6 @@ import (
 
 	"github.com/maxmcd/dag"
 )
-
-type DerivationsMap struct {
-	d    map[string]*Derivation
-	lock sync.RWMutex
-}
-
-func (dm *DerivationsMap) Load(filename string) *Derivation {
-	dm.lock.RLock()
-	defer dm.lock.RUnlock()
-	return dm.d[filename]
-}
-
-func (dm *DerivationsMap) Has(filename string) bool {
-	return dm.Load(filename) != nil
-}
-func (dm *DerivationsMap) Store(filename string, drv *Derivation) {
-	dm.lock.Lock()
-	defer dm.lock.Unlock()
-	dm.d[filename] = drv
-}
-
-func (dm *DerivationsMap) Range(cb func(map[string]*Derivation)) {
-	dm.lock.Lock()
-	cb(dm.d)
-	dm.lock.Unlock()
-}
 
 // FakeDAGRoot is used when we have multiple build outputs, or "roots" in our
 // graph so we need to tie them to a single fake root so that we still have a
@@ -53,7 +27,7 @@ func (ag AcyclicGraph) PrintDot() {
 	fmt.Println(strings.ReplaceAll(graphString, "\"[root] ", "\""))
 }
 
-func mergeGraphs(graphs ...*AcyclicGraph) *AcyclicGraph {
+func MergeGraphs(graphs ...*AcyclicGraph) *AcyclicGraph {
 	if len(graphs) == 0 {
 		return NewAcyclicGraph()
 	}
